refactor: simplificando o código para melhor legibilidade
