+++ src/pages/Home/Home.jsx
@@ -1,18 +1,12 @@
-import { useEffect, useState } from "react";
-import { toast } from "sonner";
 import { useAuth } from "../../contexts/AuthContext.jsx";
 
 import AddTask from "../../components/AddTask/AddTask";
 import { HeaderNav } from "../../components/Header/Header";
 import { SelectButton } from "../../components/Select/Select";
-import { CardTask } from "../../components/CardTasks/CardTasks.jsx";
 import { ToastPopUp } from "../../components/Toast/Toast.jsx";
 import { SearchNotes } from "../../components/SearchNotes/SearchNotes.jsx";
 import { CardTaskModal } from "../../components/Modal/CardTaskModal.jsx";
 import { Metrics } from "../../components/Metrics/index.jsx";
-
-import { getNotes, searchNotes } from "../../services/notesService.js";
-import { useDebounce } from "../../hooks/useDebounce.js";
 
 import {
 	ContainerBody,
@@ -26,91 +20,32 @@ import {
     BackgroundModal,
 	NoContent,
 } from "./Home.style";
-import { SkeletonCards } from "../../components/Skeletons/SkeletonCards/SkeletonCards.jsx";
+
+import { useNotes } from "../../hooks/useNotes.js";
+import { useModal } from "../../hooks/useModal.js";
+import { TaskList } from "../../components/TasksList/TasksList.jsx";
 
 export default function Home() {
     const { nome } = useAuth();
+	const {
+		notes,
+		handleSearch,
+		updateNote,
+		deleteNote,
+		filter,
+		setFilter,
+		isSearching,
+		loading,
+	} = useNotes();
 
-    const [modalOpen, setModalOpen] = useState(false);
-    const [currentNote, setCurrentNote] = useState(null);
-    const [notes, setNotes] = useState([]);
-	const [filter, setFilter] = useState("todas");
-    const [searchTerm, setSearchTerm] = useState("");
-    const [isSearching, setIsSearching] = useState(false);
-	const [loading,setLoading] = useState(false)
-	
-    const debouncedSearch = useDebounce(searchTerm, 500);
-
-	const applyFilter = () => {
-        if (filter === "todas") {
-            return notes;
-        }
-        return notes.filter((note) => note.status === filter);
-    };
-
-    const fetchNotes = async (searchTerm = "") => {
-		setLoading(true)
-		try {
-            if (searchTerm.trim() !== "") {
-                const response = await searchNotes(searchTerm);
-
-				if(!response.sucess){
-					toast.error(response.msg);
-				}
-
-				if(response.sucess){
-					setNotes(response.data);
-				}
-
-            } else {
-                const response = await getNotes();
-				
-				if(!response.sucess){
-					toast.error(response.msg);
-				}
-
-				if(response.sucess){
-					setNotes(response.data);
-				}
-
-            }
-        } catch (error) {
-           toast.error(error.message)
-        } finally {
-			setLoading(false)
-		}
-    };
-
-    useEffect(() => {
-        fetchNotes(debouncedSearch);
-    }, [debouncedSearch]);
-
-    const handleSearch = (term) => {
-        setSearchTerm(term);
-        setIsSearching(term.trim() !== "");
-    };
-
-    const updateNote = (updateNote) => {
-        setNotes((prevNotes) => 
-            prevNotes.map((note) =>
-                note._id === updateNote._id ? updateNote : note
-            ) 
-        )
-    };
-
-    const deleteNote = (noteId) => {
-        setNotes((prevNotes) => prevNotes.filter((note) => note._id !== noteId));
-    };
-
-    const handleCardClick = (note) => {
-        setCurrentNote(note);
-        setModalOpen(true);
-    }
-
-    const handleCloseModal = () => {
-        setCurrentNote(null);
-        setModalOpen(false);
-    }
+	const { 
+		closeModal,
+		openModal,
+		setModalOpen,
+		modalOpen,
+		currentNote,
+		setCurrentNote
+	} = useModal();
 
 	return (
 		<ContainerBody>
@@ -148,46 +83,25 @@ export default function Home() {
 					</HeaderTasks>
 
 					<ContainerTasks>
-						<ContainerCardsTasks>
-							{loading ? (
-								<SkeletonCards count={20} />
-							) : (
-								applyFilter().map((note) => (
-									<CardTask
-										key={note._id}
-										title={note.titulo}
-										description={note.descricao}
-										status={note.status}
-										onClick={() => handleCardClick(note)}
-									/>
-								))
-							)}
-						</ContainerCardsTasks>
-
+						<TaskList
+							handleCardClick={openModal}
+							loading={loading}
+							notes={notes}
+						/>
 						{currentNote && modalOpen && (
 							<>
-								<BackgroundModal onClick={handleCloseModal} />
+								<BackgroundModal onClick={closeModal} />
 								<CardTaskModal
 									onUpdate={updateNote}
 									onDelete={deleteNote}
 									noteId={currentNote._id}
 									isOpen={modalOpen}
-									onClose={handleCloseModal}
+									onClose={closeModal}
 									key={currentNote._id}
 									title={currentNote.titulo}
 									content={currentNote.descricao}
 								/>
 							</>
-						)}
-
-						{notes.length === 0 && (
-							<NoContent>
-								<img
-									src="/NoContent.png"
-									alt="Nenhuma tarefa encontrada"
-								/>
-								<h2>Adicione uma tarefa nova</h2>
-							</NoContent>
 						)}
 					</ContainerTasks>
 				</Main>
